re-import from upstream (partly for https://trac.osgeo.org/gdal/ticket/3876)
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@21231 f0d54148-0727-0410-94bb-9a71ac55c965

+++ frmts/pcidsk/sdk/segment/cpcidsk_tex.cpp
@@ -61,7 +61,7 @@ std::string CPCIDSK_TEX::ReadText()
 {
     PCIDSKBuffer seg_data;
 
-    seg_data.SetSize( GetContentSize() );
+    seg_data.SetSize( (int) GetContentSize() );
 
     ReadFromFile( seg_data.buffer, 0, seg_data.buffer_size );
